+++ backend/app/repository/genre/sql_genre_repository.py
@@ -2,9 +2,10 @@ from sqlalchemy import func, select
 from sqlalchemy.orm import Session
 
 from app.model.genre_model import Genre
+from app.repository.genre.i_genre_repository import IGenreRepository
 
 
-class SQLGenreRepository:
+class SQLGenreRepository(IGenreRepository):
     def __init__(self, db: Session):
         self.db = db
 
@@ -39,6 +40,10 @@ class SQLGenreRepository:
         self.db.commit()
         return genre
 
-    def delete(self, genre: Genre) -> None:
+    def delete(self, genre_id: int) -> None:
+        genre = self.db.get(Genre, genre_id)
+        if not genre:
+            return False
         self.db.delete(genre)
         self.db.commit()
+        return True
